Deprecate editor factories
