+++ net.sourceforge.pmd.eclipse.plugin/src/main/java/net/sourceforge/pmd/eclipse/ui/preferences/editors/MultiEnumerationEditorFactory.java
@@ -31,7 +31,10 @@ import net.sourceforge.pmd.properties.PropertySource;
  * @author Brian Remedios
  *
  *         ! PLACEHOLDER ONLY - NOT FINISHED YET !
+ * @deprecated This editor factory will be removed without replacement. This was only used for supporting the UI
+ *             of the plugin and is considered internal API now.
  */
+@Deprecated // for removal
 public final class MultiEnumerationEditorFactory extends AbstractMultiValueEditorFactory<Object> {
 
     public static final MultiEnumerationEditorFactory INSTANCE = new MultiEnumerationEditorFactory();
