+++ net.sourceforge.pmd.eclipse.plugin/src/main/java/net/sourceforge/pmd/eclipse/ui/preferences/editors/MultiStringEditorFactory.java
@@ -22,7 +22,10 @@ import net.sourceforge.pmd.properties.PropertySource;
 
 /**
  * @author Brian Remedios
+ * @deprecated This editor factory will be removed without replacement. This was only used for supporting the UI
+ *             of the plugin and is considered internal API now.
  */
+@Deprecated // for removal
 public final class MultiStringEditorFactory extends AbstractMultiValueEditorFactory<String> {
 
     public static final MultiStringEditorFactory INSTANCE = new MultiStringEditorFactory();
